+++ readme.md
@@ -84,6 +84,10 @@ Each Campground has:
 ##Auth Pt. 5 - Show/Hide Links
 * Show/hide auth links in navbar correctly
 
+##Refactor The Routes
+* Use Express router to reorganize all routes
+
+
 RESTFUL ROUTES
 
 name     url                            verb      desc
